added stopifnot function, cleaned up documentation, changed loops to lapply

+++ R/rename_files.R
@@ -1,4 +1,4 @@
-#' Rename Default File Names from Aqualog
+#' Rename default file names from Aqualog
 #'
 #' Takes the raw files downloaded from the Aqualog and renames them so file names are consistent
 #' across all run types and runs for consistent analysis and processing later.
@@ -9,7 +9,9 @@
 #' @param prjpath a string indicating the project file containing the data to process, they should not be in subfiles
 #' @export
 
-files_rename <- function(meta=meta, prjpath){
+files_rename <- function(meta, prjpath){
+  stopifnot(is.data.frame(meta) | is.character(prjpath) |file.exists(prjpath))
+
   #correct for any blanks not listed as B1
   file_correct <- list.files(prjpath)
   file_correct <- file_correct[stringr::str_detect(file_correct, ".dat")]
@@ -105,14 +107,18 @@ files_rename <- function(meta=meta, prjpath){
 #' @importFrom readxl read_xlsx
 #' @importFrom stringr str_detect
 #'
-#' @param prjpath The file path of the main file directory where data is located
+#' @param prjpath the file path of the main file directory where data is located
 #' @param meta_file file location of metadata table, used to determine how samples were run
 #' @param meta_sheet optional, sheet name of the metadata only required if metadata is an .xlsx file
-#' @param zip_files Will create a zip file in the file directory of all the raw, nonrenamed files from the Aqualog as a backup
+#' @param zip_files will create a zip file in the file directory of all the raw, non-renamed files from the Aqualog as a backup
+#' @return metadata table with an added column for the unique ID
 #' @export
 
 clean_files <- function(prjpath, meta_file, meta_sheet, zip_files=T){
-  # Load Sample Log
+  stopifnot(is.character(c(prjpath, meta_sheet, meta_file))|
+              is.logical(zip_files)| file.exists(prjpath))
+
+  #Load Sample Log
   if(stringr::str_detect(meta_file, ".xlsx")){
     meta <- readxl::read_xlsx(meta_file, sheet = meta_sheet)
   } else{
@@ -125,49 +131,53 @@ clean_files <- function(prjpath, meta_file, meta_sheet, zip_files=T){
 
   #get unique ID
   meta$unique_ID <- paste(meta$data_identifier, "_", meta$replicate_no, "_", meta$integration_time_s, "s", sep="")
+  meta <- as.data.frame(meta)
+  row.names(meta) <- meta$unique_ID
 
   #zip raw files to have a backup
-  if(zip_files == T){
+  file_to_zip <- list.files(prjpath)[stringr::str_detect(list.files(prjpath), ".dat")]
+
+  if(zip_files == T & length(file_to_zip) > 0){
     zip_name <- unlist(strsplit(prjpath, "/"))
     zip_name <- zip_name[length(zip_name)]
-    file_to_zip <- list.files(prjpath)[stringr::str_detect(list.files(prjpath), ".dat")]
     zip(paste("rawfiles_", zip_name, ".zip", sep=""), paste(file_to_zip, sep="/"))
+  }else if(length(file_to_zip) == 0 & zip_files == T){
+    warning("No raw files were found to zip.")
   }
 
   #checks to see if there are files to rename
-  stopifnot(sum(stringr::str_detect(list.files(prjpath), "SEM")) +
+  if(sum(stringr::str_detect(list.files(prjpath), "SEM")) +
               sum(stringr::str_detect(list.files(prjpath), "Waterfall")) +
               sum(stringr::str_detect(list.files(prjpath), "Abs ")) +
               sum(stringr::str_detect(list.files(prjpath), "BEM")) +
-              sum(stringr::str_detect(list.files(prjpath), "ABS")) > 0)
+              sum(stringr::str_detect(list.files(prjpath), "ABS")) > 0){
+    #rename files
+    files_rename(meta=meta, prjpath=prjpath)
 
-  #rename files
-  files_rename(meta=meta, prjpath=prjpath)
+    #identify EEMs and Absorbance files
+    files <- list.files(prjpath)
+    Abs <- files[str_detect(files, "_Abs.dat")]
+    blank <- files[str_detect(files, "_blank.dat")]
+    EEM <- files[str_detect(files, ".dat")]
+    EEM <- EEM[!(EEM %in% c(Abs, blank))]
 
-  #identify EEMs and Absorbance files
-  files <- list.files(prjpath)
-  Abs <- files[str_detect(files, "_Abs.dat")]
-  blank <- files[str_detect(files, "_blank.dat")]
-  EEM <- files[str_detect(files, ".dat")]
-  EEM <- EEM[!(EEM %in% c(Abs, blank))]
+    stopifnot(length(Abs) > 0| length(blank) > 0| length(EEM) > 0)
 
-  stopifnot(length(Abs) > 0| length(blank) > 0| length(EEM) > 0)
-  #put in files
-  dir.create(paste(prjpath, "/1_Absorbance", sep=""), showWarnings = F)
-  dir.create(paste(prjpath, "/2_Blanks", sep=""), showWarnings = F)
-  dir.create(paste(prjpath, "/3_Samples", sep=""), showWarnings = F)
+    create_files(prjpath)
 
-  file.copy(paste(prjpath, "/", Abs, sep=""), paste(prjpath, "/1_Absorbance/",sep=""))
-  file.copy(paste(prjpath, "/", blank, sep=""), paste(prjpath, "/2_Blanks/",sep=""))
-  file.copy(paste(prjpath, "/", EEM, sep=""), paste(prjpath, "/3_Samples/",sep=""))
+    file.copy(paste(prjpath, "/", Abs, sep=""), paste(prjpath, "/1_Absorbance/",sep=""))
+    file.copy(paste(prjpath, "/", blank, sep=""), paste(prjpath, "/2_Blanks/",sep=""))
+    file.copy(paste(prjpath, "/", EEM, sep=""), paste(prjpath, "/3_Samples/",sep=""))
 
-  file.remove(paste(prjpath, "/", Abs, sep=""))
-  file.remove(paste(prjpath, "/", blank, sep=""))
-  file.remove(paste(prjpath, "/", EEM, sep=""))
+    file.remove(paste(prjpath, "/", Abs, sep=""))
+    file.remove(paste(prjpath, "/", blank, sep=""))
+    file.remove(paste(prjpath, "/", EEM, sep=""))
 
-  #check you have absorbance, blank, and sample data for each sample
-  if(length(EEM) != length(blank) & length(EEM) != length(Abs)){
-    stop("Warning: Your samples aren't matched, you're missing absorbance or EEM's data")
+    #check you have absorbance, blank, and sample data for each sample
+    if(length(EEM) != length(blank) & length(EEM) != length(Abs)){
+      stop("Warning: Your samples aren't matched, you're missing absorbance or EEM's data")
+    }
   }
 
+  return(meta)
 }
